Refactor EditPersonDescriptor to a separate class
