+++ src/main/java/seedu/address/logic/parser/EditPersonDescriptor.java
@@ -1,0 +1,141 @@
+package seedu.address.logic.parser;
+
+import seedu.address.commons.util.CollectionUtil;
+import seedu.address.model.person.*;
+import seedu.address.model.tag.Tag;
+
+import java.time.LocalDateTime;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
+
+/**
+ * Stores the details to edit the person with. Each non-empty field value will replace the
+ * corresponding field value of the person.
+ */
+public class EditPersonDescriptor {
+    private Name name;
+    private Phone phone;
+    private Email email;
+    private Address address;
+    private Set<Tag> tags;
+    private LocalDateTime localdateTime;
+
+    public EditPersonDescriptor() {}
+
+    /**
+     * Copy constructor.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public EditPersonDescriptor(EditPersonDescriptor toCopy) {
+        setName(toCopy.name);
+        setPhone(toCopy.phone);
+        setEmail(toCopy.email);
+        setAddress(toCopy.address);
+        setTags(toCopy.tags);
+        setDate(toCopy.localdateTime);
+    }
+
+    /**
+     * Returns true if at least one field is edited.
+     */
+    public boolean isAnyFieldEdited() {
+        return CollectionUtil.isAnyNonNull(name, phone, email, address, tags);
+    }
+
+    public void setName(Name name) {
+        this.name = name;
+    }
+
+    public Optional<Name> getName() {
+        return Optional.ofNullable(name);
+    }
+
+    public void setPhone(Phone phone) {
+        this.phone = phone;
+    }
+
+    public Optional<Phone> getPhone() {
+        return Optional.ofNullable(phone);
+    }
+
+    public void setEmail(Email email) {
+        this.email = email;
+    }
+
+    public Optional<Email> getEmail() {
+        return Optional.ofNullable(email);
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+
+    public void setDate(LocalDateTime localdateTime) {
+        this.localdateTime= localdateTime;
+    }
+
+    public Optional<LocalDateTime> getDate() {
+        return Optional.ofNullable(localdateTime);
+    }
+
+    public Optional<Address> getAddress() {
+        return Optional.ofNullable(address);
+    }
+
+    /**
+     * Sets {@code tags} to this object's {@code tags}.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public void setTags(Set<Tag> tags) {
+        this.tags = (tags != null) ? new HashSet<>(tags) : null;
+    }
+
+    /**
+     * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
+     * if modification is attempted.
+     * Returns {@code Optional#empty()} if {@code tags} is null.
+     */
+    public Optional<Set<Tag>> getTags() {
+        return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
+    }
+
+    /**
+     * Creates and returns a {@code Person} with the details of {@code personToEdit}
+     * edited with {@code editPersonDescriptor}.
+     */
+    public static Person createEditedPerson(Person personToEdit, EditPersonDescriptor editPersonDescriptor) {
+        assert personToEdit != null;
+
+        Name updatedName = editPersonDescriptor.getName().orElse(personToEdit.getName());
+        Phone updatedPhone = editPersonDescriptor.getPhone().orElse(personToEdit.getPhone());
+        Email updatedEmail = editPersonDescriptor.getEmail().orElse(personToEdit.getEmail());
+        Address updatedAddress = editPersonDescriptor.getAddress().orElse(personToEdit.getAddress());
+        Set<Tag> updatedTags = editPersonDescriptor.getTags().orElse(personToEdit.getTags());
+
+        return new Person(updatedName, updatedPhone, updatedEmail, updatedAddress, updatedTags);
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        // short circuit if same object
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof EditPersonDescriptor)) {
+            return false;
+        }
+
+        // state check
+        EditPersonDescriptor e = (EditPersonDescriptor) other;
+
+        return getName().equals(e.getName())
+                && getPhone().equals(e.getPhone())
+                && getEmail().equals(e.getEmail())
+                && getAddress().equals(e.getAddress())
+                && getTags().equals(e.getTags());
+    }
+}
